fix(harness): lower plugin import realm boundary
Refs #492
